improvment of cache sweeper

+++ config/initializers/directory_sweeper.rb
@@ -3,15 +3,19 @@ module ActionController
 
     class Sweeper
 
-      def sweep_directory(dirs)
+      PROTECTED_DIRECTORIES = ["images", "javascripts", "stylesheets", "system"]
+
+      include ActionController::UrlWriter
+
+      def sweep_directory(*dirs)
         dirs.each do |dir|
-          cache_dir = RAILS_ROOT + "/public/#{dir}"
+          next if PROTECTED_DIRECTORIES.include?(dir[/\w+/]) or !dir[/\w+/]
+          cache_dir = RAILS_ROOT + "/public/#{dir[/\w+/]}"
           FileUtils.rm_r(Dir.glob(cache_dir+"/*")) rescue Errno::ENOENT
           RAILS_DEFAULT_LOGGER.info("Cache directory '#{cache_dir}' fully swept.")
         end
       end
-
     end
-
   end
 end
+
